requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
h11==0.13.0
idna==3.3
outcome==1.2.0
PySocks==1.7.1
python-slugify==6.1.2
requests==2.28.1
selenium==4.4.3
sniffio==1.2.0
sortedcontainers==2.4.0
text-unidecode==1.3
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
